+++ tests/test_server.py
@@ -1,7 +1,8 @@
-from grpcalchemy.blueprint import Blueprint
+from grpcalchemy.blueprint import Blueprint, Context
 from grpcalchemy.client import Client
 from grpcalchemy.orm import Message, StringField
 from grpcalchemy.server import Server
+
 from .test_grpcalchemy import TestGrpcalchemy
 
 
@@ -15,7 +16,8 @@ class TestServer(TestGrpcalchemy):
         self.test_blueprint = Blueprint("test_blueprint")
 
         @self.test_blueprint.register
-        def test_message(request: TestMessage, context) -> TestMessage:
+        def test_message(request: TestMessage,
+                         context: Context) -> TestMessage:
             return TestMessage(test_name=request.test_name)
 
         self.test_message = test_message
@@ -31,6 +33,7 @@ class TestServer(TestGrpcalchemy):
         test_name = "Hello World!"
         with Client("localhost:50051") as client:
             client.register(self.test_blueprint)
-            response = client.test_blueprint(self.test_message,
-                                             self.Message(test_name=test_name))
+            response = client.test_blueprint(
+                rpc=self.test_message,
+                message=self.Message(test_name=test_name))
             self.assertEqual(test_name, response.test_name)